Version 3.3 Fixed silent tone bug for Lollipop and other bugs. Multiline text /r /n handling. Android time accessible on Arduino. Debug messages as toasts. Changed create button. Slider values scalable. GUI multi touch. Hex and ASCII output of received Bluetooth data at log level verbose.

+++ arduino/libraries/BlueDisplay/BDButton.h
@@ -76,7 +76,7 @@ public:
     bool operator!=(const BDButton &aButton);
 
     void init(uint16_t aPositionX, uint16_t aPositionY, uint16_t aWidthX, uint16_t aHeightY, Color_t aButtonColor,
-            const char * aCaption, uint8_t aCaptionSize, uint8_t aFlags, int16_t aValue,
+            const char * aCaption, uint16_t aCaptionSize, uint8_t aFlags, int16_t aValue,
             void (*aOnTouchHandler)(BDButton*, int16_t));
 
     void drawButton(void);
